moves validate_objects and regex_matcher into libs from tools
