+++ openvair/modules/image/entrypoints/crud.py
@@ -12,12 +12,12 @@ Classes:
 from typing import Dict, List, Optional, cast
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import validate_objects
 from openvair.modules.image.config import (
     PERMITTED_EXTENSIONS,
     API_SERVICE_LAYER_QUEUE_NAME,
 )
 from openvair.modules.image.entrypoints import schemas
+from openvair.libs.validation.validators import validate_objects
 from openvair.modules.image.service_layer import services
 from openvair.libs.messaging.messaging_agents import MessagingClient
 from openvair.modules.image.entrypoints.exceptions import (
