+++ openvair/modules/notification/entrypoints/api.py
@@ -22,12 +22,11 @@ from starlette.responses import JSONResponse
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import regex_matcher, get_current_user
+from openvair.modules.tools.utils import get_current_user
 from openvair.modules.notification.entrypoints import schemas
 from openvair.modules.notification.entrypoints.crud import NotificationCrud
 
 LOG = get_logger(__name__)
-UUID_REGEX = regex_matcher('uuid4')
 
 router = APIRouter(
     prefix='/notifications',
