+++ openvair/modules/tools/utils.py
@@ -17,7 +17,6 @@ Functions:
     lip_scan: Performs LIP (Loop Initialization Protocol) scan.
     get_size: Returns the size of a file.
     validate_objects: Validates a list of objects against a Pydantic schema.
-    regex_matcher: Returns regex patterns for various types of values.
     write_yaml_file: Writes data to a YAML file.
     read_yaml_file: Reads data from a YAML file.
     synchronized_session: Context manager for safely executing database
@@ -31,7 +30,6 @@ from typing import (
     Any,
     Dict,
     List,
-    Type,
     Union,
     Optional,
     Generator,
@@ -44,7 +42,6 @@ import jwt
 import yaml
 import xmltodict
 from fastapi import Depends, HTTPException, security
-from pydantic import BaseModel, TypeAdapter, ValidationError
 from sqlalchemy.exc import OperationalError
 
 from openvair import config
@@ -255,7 +252,7 @@ def get_block_devices_info() -> List[Dict[str, str]]:
         '--json',
         params=ExecuteParams(  # noqa: S604
             shell=True,
-        )
+        ),
     )
     result: List[Dict[str, str]] = json.loads(res.stdout)['blockdevices']
     return result
@@ -407,130 +404,6 @@ def get_size(file_path: str) -> int:
         int: The size of the file in bytes.
     """
     return Path(file_path).stat().st_size
-
-
-def _create_corrupted_data(
-    pydantic_schema: Type[BaseModel],
-    _object: Dict[str, Any],
-) -> Dict[str, Any]:
-    corrupted_data: Dict[str, Any] = {}
-    for (
-        field_name,
-        field_info,
-    ) in pydantic_schema.model_fields.items():
-        if field_name == 'id':
-            corrupted_data[field_name] = _object.get('id')
-        elif field_name == 'status':
-            corrupted_data[field_name] = 'corrupted object'
-        else:
-            adapter: TypeAdapter[Any] = TypeAdapter(field_info.annotation)
-            try:
-                corrupted_data[field_name] = adapter.validate_python(
-                    None, strict=False
-                )
-            except ValidationError:
-                corrupted_data[field_name] = None
-    return corrupted_data
-
-
-def validate_objects(
-    objects: List[Dict[str, Any]],
-    pydantic_schema: Type[BaseModel],
-    *,
-    skip_corrupted_object: bool = True,
-) -> List[BaseModel]:
-    """Validates a list of objects against a Pydantic schema
-
-    Ensures that all returned objects are valid instances of the schema.
-
-    This function processes a list of dictionary-based objects, attempting to
-    validate each object against the provided Pydantic schema. If an object
-    fails validation, it can either be replaced with a "corrupted object"
-    (containing default values that satisfy the schema) or raise an exception,
-    depending on the `skip_corrupted_object` parameter.
-
-    The function guarantees that all returned objects conform to the schema,
-    making it suitable for use in scenarios where subsequent processing (e.g.,
-    API responses in FastAPI) requires fully valid Pydantic models.
-
-    Args:
-        objects (List[Dict[str, Any]]):
-            A list of dictionaries representing objects to be validated.
-        pydantic_schema (Type[BaseModel]):
-            The Pydantic schema against which each object will be validated.
-        skip_corrupted_object (bool, optional):
-            If True (default), objects that fail validation are replaced with
-            a "corrupted object" containing default values.
-            If False, the function raises a `ValidationError` upon encountering
-            an invalid object.
-
-    Returns:
-        List[BaseModel]:
-            A list of validated Pydantic objects, where all elements conform
-            to the specified schema. If `skip_corrupted_object=True`,
-            invalid objects are replaced with valid "corrupted" versions.
-
-    Raises:
-        ValidationError:
-            If `skip_corrupted_object=False` and an object fails validation,
-            an exception is raised instead of replacing it.
-
-    Example:
-        >>> from pydantic import BaseModel
-        >>> from typing import List
-        >>> class UserModel(BaseModel):
-        ...     id: int
-        ...     name: str
-        ...     status: str = 'active'
-        >>> objects = [
-        ...     {'id': 1, 'name': 'Alice'},
-        ...     {'id': 2, 'name': 123},  # Invalid: name should be str
-        ...     {'id': '3'},  # Missing name (required field)
-        ... ]
-        >>> valid_users = validate_objects(objects, UserModel)
-        >>> for user in valid_users:
-        ...     print(user)
-        UserModel(id=1, name='Alice', status='active')
-        UserModel(id=2, name='', status='corrupted object') # Replaced invalid entry
-        UserModel(id=3, name='', status='corrupted object') # Replaced invalid entry
-
-    """  # noqa: E501
-    result: List[BaseModel] = []
-    for _object in objects:
-        try:
-            validated_object = pydantic_schema.model_validate(_object)
-            result.append(validated_object)
-        except ValidationError as err:
-            message = (
-                f'Validation error: {err}\nWhile validating object: {_object} '
-                f'with schema: {pydantic_schema.__name__}'
-            )
-            LOG.warning(message)
-            if skip_corrupted_object:
-                corrupted_object: BaseModel = pydantic_schema.model_construct(
-                    **_create_corrupted_data(pydantic_schema, _object)
-                )
-                result.append(corrupted_object)
-            else:
-                LOG.error(message)
-                raise
-    return result
-
-def regex_matcher(value: str) -> str:
-    """Returns regex patterns for various types of values.
-
-    Args:
-        value (str): The type of value to match (e.g., 'mac_address', 'uuid4').
-
-    Returns:
-        str: The regex pattern for the specified value type.
-    """
-    regex_dict = {
-        'mac_address': r'^([0-9A-F]{2}:){5}[0-9A-F]{2}',
-        'uuid4': r'[a-f0-9]{8}-?[a-f0-9]{4}-?4[a-f0-9]{3}-?[89ab][a-f0-9]{3}-?[a-f0-9]{12}$',  # noqa: E501 because its regex pattern
-        'special_characters': r'(?=.*[ -\/:-@\[-\`{-~]{1,})',
-    }
-    return regex_dict[value]
 
 
 def write_yaml_file(file_path: str, data: Dict) -> None:
