+++ openvair/modules/volume/entrypoints/api.py
@@ -27,11 +27,8 @@ from fastapi_pagination import Page, paginate
 from starlette.concurrency import run_in_threadpool
 
 from openvair.libs.log import get_logger
-from openvair.modules.tools.utils import (
-    regex_matcher,
-    get_current_user,
-    validate_objects,
-)
+from openvair.modules.tools.utils import get_current_user
+from openvair.libs.validation.validators import regex_matcher, validate_objects
 from openvair.modules.volume.entrypoints import schemas
 from openvair.modules.volume.entrypoints.crud import VolumeCrud
